VAPI-3948 Widen sdpRevision from Int to Long
The gateway sends sdpRevision values large enough to overflow a 32-bit
Int - observed as millisecond timestamps. kotlinx.serialization throws
decoding the sdpOffer notification whenever this happens, and the
exception is swallowed by the generic event-handler error path, so the
client silently never negotiates the subscribe peer connection at all.

Reproduced live: a real outbound PSTN call was answered on the far end,
but the WebRTC subscribe leg never connected because the SDP offer
notification failed to deserialize.

Widen SDPOfferNotification.sdpRevision, PeerConnectionManager's
subscribeSdpRevision field and handleSubscribeSdpOffer parameter, and
the matching PeerConnectionManagerInterface signature to Long.

+++ bandwidthrtc/src/main/kotlin/com/bandwidth/rtc/webrtc/PeerConnectionManager.kt
@@ -40,7 +40,7 @@ class PeerConnectionManager(
     private val publishedStreams = ConcurrentHashMap<String, MediaStream>()
     private val publishedAudioSources = ConcurrentHashMap<String, AudioSource>()
     private val subscribedTrackMetadata = ConcurrentHashMap<String, TrackMetadata>()
-    var subscribeSdpRevision: Int = 0
+    var subscribeSdpRevision: Long = 0
         private set
 
     override var onStreamAvailable: ((MediaStream, List<MediaType>, TrackMetadata?) -> Unit)? = null
@@ -246,17 +246,17 @@ class PeerConnectionManager(
 
     override suspend fun handleSubscribeSdpOffer(
         sdpOffer: String,
-        sdpRevision: Int?,
+        sdpRevision: Long?,
         metadata: Map<String, TrackMetadata>?
     ): String {
         val effectiveRevision = sdpRevision ?: (subscribeSdpRevision + 1)
 
-        if (effectiveRevision <= subscribeSdpRevision && subscribeSdpRevision != 0) {
+        if (effectiveRevision <= subscribeSdpRevision && subscribeSdpRevision != 0L) {
             log.warn("Rejecting stale SDP offer (revision $effectiveRevision <= $subscribeSdpRevision)")
             throw BandwidthRTCError.SdpNegotiationFailed("Stale SDP offer")
         }
 
-        if (subscribeSdpRevision == 0) {
+        if (subscribeSdpRevision == 0L) {
             log.debug("Accepting first subscribe SDP offer (revision 0→$effectiveRevision)")
         }
